Settable grpc logger (#402)
* SettableLoggerV2 to dynamically change grpc logger.

The go-grpc assumes that logger can be only configured once as the `SetLoggerV2`
method is:
```Not mutex-protected, should be called before any gRPC functions.``

This is insufficient in case of e.g. "testing" scenarios where
loggers might need to be supplied on per-test-case basis.

* Integrate ZAP with SettableLoggerV2.

The settable_test.go is a good example of integration of
testing harness with grpclogging using zaptest,
that is canonical usecase for this infrastructure.