Fix memory leakage in kit.PayloadUnaryServerInterceptor (#501)
* Fix memory leakage in kit.PayloadUnaryServerInterceptor


Closes issue #498

A wrong assign creates multiple object for GC, and with every call appends new objects, so GC have to walk through thousands of objects. This causes high increasing memory and CPU usage. We saw in our services that it leads to CPU throttling and increasing of a memory usage.

* Fix logging in PayloadUnaryServerInterceptor

Issue #498