fix: Refactor logger initialization in example_test.go (#580)
This commit refactors the logger initialization in the example_test.go files
for the logr, logrus, and zap interceptors. In each file, the InterceptorLogger
function was modified to create a new logger instance using the original logger
and additional fields passed in as arguments, rather than modifying the original
logger instance. This ensures that subsequent logs using the original logger are not
affected by the additional fields.

The changes were made to the following files:
- interceptors/logging/examples/logr/example_test.go
- interceptors/logging/examples/logrus/example_test.go
- interceptors/logging/examples/zap/example_test.go

Signed-off-by: aimuz <mr.imuz@gmail.com>